Permissions check

+++ events/message.js
@@ -65,7 +65,9 @@ exports.run = (client, msg) => {
 		const messagechannel = client.channels.get(botconfig.activitychannel);
 		messagechannel.send({ embed: activityembed });
 	}
-		cmd.run(client, msg, args);
+	if (cmd.help.botpermissions.every(perm => msg.guild.me.hasPermission(perm)) === false) return msg.channel.send(`It looks like the bot hasn't enough permissions to execute this command! (Required permissions: ${cmd.help.botpermissions.join(', ')})`);
+	if (cmd.conf.userpermissions.every(perm => msg.member.hasPermission(perm)) === false) return msg.channel.send(`It looks like you haven't enough permissions to execute this command! (Required permissions: ${cmd.conf.userpermissions.join(', ')})`);
+	cmd.run(client, msg, args);
 		if (tableload.commanddel === 'true') {
 			msg.delete();
 		}
